Rename package to a more meaningful name
